requirements.txt
asgiref==3.8.1
Django==4.1.13
django-filter==24.2
djangorestframework==3.15.2
djongo==1.3.6
dnspython==2.6.1
Markdown==3.6
pymongo==4.8.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.2.4
typing_extensions==4.12.2
